Break bug modal out & enforce GH issue max url length
